// frontend/src/generated/openapi/model/paginaLibros.ts
/**
 * amoxtli
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InfoBasicaLibro } from './infoBasicaLibro';


export interface PaginaLibros { 
    /**
     * Número de página siguiente
     */
    pagSiguiente: number;
    /**
     * Número de página anterior
     */
    pagAnterior: number;
    /**
     * Número de resultados por página
     */
    resultados: number;
    libros: Array<InfoBasicaLibro>;
}